moved 'render ssh' to 'render services ssh' + moves 'render ssh <serviceId> to positional'; fixes #18 + #19
